Improved Travis build information stage
